Update link selector to format links
The link selector is working correctly, as it's currently set to type="url" however Notion has this set to type="text" meaning browser validation isn't run on the input to check it is a valid URL. This means in notion typing google.com will link to http://google.com and allow the domain to handle http > https redirection. I have added a snippet to the form submission which will trim the string and then append http:// if it doesn't start with it already as well as changing the type to text on the input field.